+++ services/web/mgmt/2019-08-01/web/domains.go
@@ -255,7 +255,7 @@ func (client DomainsClient) CreateOrUpdateResponder(resp *http.Response) (result
 }
 
 // CreateOrUpdateOwnershipIdentifier description for Creates an ownership identifier for a domain or updates identifier
-// details for an existing identifer
+// details for an existing identifier
 // Parameters:
 // resourceGroupName - name of the resource group to which the resource belongs.
 // domainName - name of domain.
@@ -1411,7 +1411,7 @@ func (client DomainsClient) UpdateResponder(resp *http.Response) (result Domain,
 }
 
 // UpdateOwnershipIdentifier description for Creates an ownership identifier for a domain or updates identifier details
-// for an existing identifer
+// for an existing identifier
 // Parameters:
 // resourceGroupName - name of the resource group to which the resource belongs.
 // domainName - name of domain.