+++ services/web/mgmt/2020-06-01/web/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. SiteConfig.AcrUseManagedIdentityCreds
+1. SiteConfig.AcrUserManagedIdentityID