+++ services/web/mgmt/2021-03-01/web/CHANGELOG.md
@@ -1,2 +1,18 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. AppsClient.CreateOneDeployOperation(context.Context, string, string) (SetObject, error)
+1. AppsClient.CreateOneDeployOperationPreparer(context.Context, string, string) (*http.Request, error)
+1. AppsClient.CreateOneDeployOperationResponder(*http.Response) (SetObject, error)
+1. AppsClient.CreateOneDeployOperationSender(*http.Request) (*http.Response, error)
+1. AppsClient.GetAuthSettingsV2WithoutSecretsSlot(context.Context, string, string, string) (SiteAuthSettingsV2, error)
+1. AppsClient.GetAuthSettingsV2WithoutSecretsSlotPreparer(context.Context, string, string, string) (*http.Request, error)
+1. AppsClient.GetAuthSettingsV2WithoutSecretsSlotResponder(*http.Response) (SiteAuthSettingsV2, error)
+1. AppsClient.GetAuthSettingsV2WithoutSecretsSlotSender(*http.Request) (*http.Response, error)
+1. AppsClient.GetOneDeployStatus(context.Context, string, string) (SetObject, error)
+1. AppsClient.GetOneDeployStatusPreparer(context.Context, string, string) (*http.Request, error)
+1. AppsClient.GetOneDeployStatusResponder(*http.Response) (SetObject, error)
+1. AppsClient.GetOneDeployStatusSender(*http.Request) (*http.Response, error)